requirements.txt
Django==3.1
psycopg2-binary==2.8.6
wheel
Pillow==7.1
djangorestframework==3.12
moesifdjango
django-filter
